feat: dataモジュールに移動
