+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/account/db/AccountRelation.kt
@@ -4,7 +4,7 @@ import androidx.room.DatabaseView
 import androidx.room.Embedded
 import androidx.room.Ignore
 import androidx.room.Relation
-import net.pantasystem.milktea.model.account.page.PageRecord
+import net.pantasystem.milktea.data.infrastructure.account.page.db.PageRecord
 
 @DatabaseView
 class AccountRelation{
