+++ modules/model/src/main/java/net/pantasystem/milktea/model/account/page/Page.kt
@@ -1,7 +1,7 @@
 package net.pantasystem.milktea.model.account.page
 
 import android.os.Parcelable
-import androidx.room.*
+import androidx.room.Ignore
 import kotlinx.parcelize.Parcelize
 import java.io.Serializable
 
@@ -35,39 +35,3 @@ data class Page(
     }
 }
 
-@Entity(
-    tableName = "page_table",
-    indices = [Index("weight"), Index("accountId")]
-)
-data class PageRecord(
-    var accountId: Long,
-    val title: String,
-    var weight: Int,
-    @Embedded val pageParams: PageParams,
-    @PrimaryKey(autoGenerate = true) var pageId: Long
-) {
-    constructor(accountId: Long, title: String, weight: Int, pageable: Pageable, pageId: Long = 0)
-            : this(accountId, title, weight, pageable.toParams(), pageId)
-
-    companion object {
-        fun from(page: Page): PageRecord {
-            return PageRecord(
-                accountId = page.accountId,
-                title = page.title,
-                weight = page.weight,
-                pageParams = page.pageParams,
-                pageId = page.pageId
-            )
-        }
-    }
-
-    fun toPage(): Page {
-        return Page(
-            accountId = accountId,
-            title = title,
-            weight = weight,
-            pageParams = pageParams,
-            pageId = pageId
-        )
-    }
-}
